requirements.txt
fastapi
uvicorn
gradio
chromadb
openai
pydantic
requests
python-dotenv
crewai>=0.50.0
pymongo